ctrl-z puts background process into the foreground
Idiom in vim that I've been using is hitting ctrl-z to drop back into
the shell and work on the command line. Quite a while ago I switched to
vim mode in my shell but didn't understand how this broke my setup.

It turns out you have to be really specific about the mode that you're
in, something that the built in emacs mode doens't care about.

To fix this, use bindkey -M [namespace]