refactor: Prefer origin/HEAD for git tools
I'm in a few codebases that use origin/main and others that use
origin/master. To standardise across htem, I was using an origin-head,
however that's clunky because I need to execute it in a sub-command to
use the output for other git commands.

Looking it it closer, I just have to start getting comfortable with
upstream's default branch being origin/HEAD.